done reset password done

+++ controller/courses.js
@@ -61,14 +61,20 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 
 exports.AddCourse = asyncHandler(async (req, res, next) => {
 
-    req.body.bootcamp = req.params.bootcampId;
+    req.body.bootcamp = req.params.bootCampId;
+    req.body.user =  req.user.id;
 
-    const bootcamp = await Bootcamp.findById(req.params.bootcampId);
+    const bootcamp = await Bootcamp.findById(req.params.bootCampId);
 
     if(!bootcamp){
         return next(new ErrorResponse(`No BootCamp with this  id of ${req.params.bootcampId}`), 404);
     };
-    
+
+      // Make sure User is BootCamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== "admin"){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootCamp ${bootcamp._id}`, 404));
+    }
+
     const course = await Course.create(req.body);
 
     res.status(200).json({
@@ -91,6 +97,11 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
     if(!course){
         return next(new ErrorResponse(`No Bootcamp with this  id of ${req.params.id}`), 404);
     };
+
+    // Make sure User is Course owner
+    if(course.user.toString() !== req.user.id && req.user.role !== "admin"){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to Update the Course ${course._id}`, 404));
+    }
     
     course = await Course.findByIdAndUpdate(req.params.id , req.body, {
         new : true,
@@ -116,6 +127,11 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
     if(!course){
         return next(new ErrorResponse(`No Bootcamp with this  id of ${req.params.id}`), 404);
     };
+
+    // Make sure User is Course owner
+    if(course.user.toString() !== req.user.id && req.user.role !== "admin"){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to Delete the Course ${course._id}`, 404));
+    }
     
     await course.deleteOne();
